+++ R/jags-model.R
@@ -98,7 +98,7 @@
 #' 
 #' model <- jags_model("
 #' model { 
-#'  bLambda ~ dlnorm(0,10^-2) 
+#'  bLambda ~ dlnorm(0,10^-2) # $\\lamda$
 #'  for (i in 1:length(x)) { 
 #'    x[i]~dpois(bLambda) 
 #'  }