+++ inst/tests/test-coef.R
@@ -3,8 +3,8 @@ context("coef")
 test_that("coef level = no", {
   
   model <- jags_model(" model { 
-                      bLambda ~ dlnorm(0,10^-2) 
-                      bLambda2 ~ dlnorm(0,10^-2) 
+                      bLambda ~ dlnorm(0,10^-2) #  $\\lambda$
+                      bLambda2 ~ dlnorm(0,10^-2) # # $\\lambda_2$
                       for (i in 1:length(x)) { 
                       x[i]~dpois(bLambda) 
                       } 
@@ -17,6 +17,11 @@ test_that("coef level = no", {
   expect_is(x, "data.frame")
   expect_identical(colnames(x), c("estimate", "lower", "upper", "sd", "error", "significance"))
   expect_identical(rownames(x), c("bLambda", "bLambda2"))
+
+  x <- coef(analysis, latex = TRUE)
+  expect_is(x, "data.frame")
+  expect_identical(colnames(x), c("estimate", "lower", "upper", "sd", "error", "significance"))
+  expect_identical(rownames(x), c("$\\lambda$", "$\\lambda_2$"))
 
   x <- coef(analysis, parm = "bLambda", level = "no")
   expect_is(x, "data.frame")